edit question functionality finished for home screen
