+++ src/components/button/goBackButton/index.ts
@@ -1,0 +1,1 @@
+export * from "./GoBackButton";
