--- src/components/button/routeButton/index.ts
@@ -1,1 +1,0 @@
-export * from "./RouteButton";
