+++ src/components/richText/editor/RichTextEditor.tsx
@@ -5,144 +5,17 @@ import { ContentEditable } from "@lexical/react/LexicalContentEditable";
 import { HistoryPlugin } from "@lexical/react/LexicalHistoryPlugin";
 import { LexicalErrorBoundary } from "@lexical/react/LexicalErrorBoundary";
 import { OnChangePlugin } from "@lexical/react/LexicalOnChangePlugin";
-
-import {
-    $isTextNode,
-    DOMConversionMap,
-    DOMExportOutput,
-    DOMExportOutputMap,
-    EditorState,
-    isHTMLElement,
-    Klass,
-    LexicalEditor,
-    LexicalNode,
-    ParagraphNode,
-    TextNode,
-} from "lexical";
+import { ToolbarPlugin } from "../plugins";
+import { EditorState, ParagraphNode, TextNode } from "lexical";
 import styles from "../RichTextShared.module.css";
-import { ToolbarPlugin } from "../";
+import utils from "./EditorUtils";
 import { CSSProperties, Dispatch, SetStateAction } from "react";
 import exampleTheme from "../exampleTheme";
 
-const MIN_ALLOWED_FONT_SIZE = 8;
-const MAX_ALLOWED_FONT_SIZE = 72;
-
-const parseAllowedFontSize = (input: string): string => {
-    const match = /^(\d+(?:\.\d+)?)px$/.exec(input);
-    if (match) {
-        const n = Number(match[1]);
-        if (n >= MIN_ALLOWED_FONT_SIZE && n <= MAX_ALLOWED_FONT_SIZE) {
-            return input;
-        }
-    }
-    return "";
-};
-
-function parseAllowedColor(input: string) {
-    return /^rgb\(\d+, \d+, \d+\)$/.test(input) ? input : "";
-}
-
-const removeStylesExportDOM = (
-    editor: LexicalEditor,
-    target: LexicalNode,
-): DOMExportOutput => {
-    const output = target.exportDOM(editor);
-    if (output && isHTMLElement(output.element)) {
-        // Remove all inline styles and classes if the element is an HTMLElement
-        // Children are checked as well since TextNode can be nested
-        // in i, b, and strong tags.
-        for (const el of [
-            output.element,
-            ...output.element.querySelectorAll('[style],[class],[dir="ltr"]'),
-        ]) {
-            el.removeAttribute("class");
-            el.removeAttribute("style");
-            if (el.getAttribute("dir") === "ltr") {
-                el.removeAttribute("dir");
-            }
-        }
-    }
-    return output;
-};
-
-const exportMap: DOMExportOutputMap = new Map<
-    Klass<LexicalNode>,
-    (editor: LexicalEditor, target: LexicalNode) => DOMExportOutput
->([
-    [ParagraphNode, removeStylesExportDOM],
-    [TextNode, removeStylesExportDOM],
-]);
-
-const getExtraStyles = (element: HTMLElement): string => {
-    // Parse styles from pasted input, but only if they match exactly the
-    // sort of styles that would be produced by exportDOM
-    let extraStyles = "";
-    const fontSize = parseAllowedFontSize(element.style.fontSize);
-    const backgroundColor = parseAllowedColor(element.style.backgroundColor);
-    const color = parseAllowedColor(element.style.color);
-    if (fontSize !== "" && fontSize !== "15px") {
-        extraStyles += `font-size: ${fontSize};`;
-    }
-    if (backgroundColor !== "" && backgroundColor !== "rgb(255, 255, 255)") {
-        extraStyles += `background-color: ${backgroundColor};`;
-    }
-    if (color !== "" && color !== "rgb(0, 0, 0)") {
-        extraStyles += `color: ${color};`;
-    }
-    return extraStyles;
-};
-
-const constructImportMap = (): DOMConversionMap => {
-    const importMap: DOMConversionMap = {};
-
-    // Wrap all TextNode importers with a function that also imports
-    // the custom styles implemented by the playground
-    for (const [tag, fn] of Object.entries(TextNode.importDOM() ?? {})) {
-        importMap[tag] = importNode => {
-            const importer = fn(importNode);
-            if (!importer) {
-                return null;
-            }
-            return {
-                ...importer,
-                conversion: element => {
-                    const output = importer.conversion(element);
-                    if (
-                        output?.forChild === undefined ||
-                        output.after !== undefined ||
-                        output.node !== null
-                    ) {
-                        return output;
-                    }
-                    const extraStyles = getExtraStyles(element);
-                    if (extraStyles) {
-                        const { forChild } = output;
-                        return {
-                            ...output,
-                            forChild: (child, parent) => {
-                                const textNode = forChild(child, parent);
-                                if ($isTextNode(textNode)) {
-                                    textNode.setStyle(
-                                        textNode.getStyle() + extraStyles,
-                                    );
-                                }
-                                return textNode;
-                            },
-                        };
-                    }
-                    return output;
-                },
-            };
-        };
-    }
-
-    return importMap;
-};
-
 const editorConfig = {
     html: {
-        export: exportMap,
-        import: constructImportMap(),
+        export: utils.exportMap,
+        import: utils.constructImportMap(),
     },
     namespace: "MyEditor",
     nodes: [ParagraphNode, TextNode],
