+++ src/components/richText/editor/index.ts
@@ -1,1 +1,2 @@
 export * from "./RichTextEditor";
+export * from "./EditorUtils";
