+++ src/components/tabs/Tabs.tsx
@@ -20,6 +20,7 @@ interface ITabsProps {
     tabBtnsContainerStyle?: CSSProperties;
     containerStyle?: CSSProperties;
     collapseWidth?: number;
+    startingTabIdx?: number;
 }
 
 //You can optionally pass styles to most elements of this component. But most
@@ -45,6 +46,11 @@ export function Tabs({
     const [activeTab, setActiveTab] = useState<number>(tabsWithIdx[0].idx);
     const [contentTopAttribute, setContentTopAttribute] = useState(0);
     const isViewportSmall = useMediaQuery(`(max-width: ${collapseWidth}px`);
+    const tab = tabsWithIdx.find(t => t.idx === activeTab);
+    if (!tab) {
+        throw new Error("Tabs components cant find the tab index");
+    }
+    const { content, contentContainerStyle } = tab;
 
     const handleTabClick = (idx: number) => {
         setActiveTab(idx);
@@ -82,20 +88,15 @@ export function Tabs({
                 )}
             </div>
 
-            {tabsWithIdx.map(({ content, idx, contentContainerStyle }) => (
-                <div
-                    style={{
-                        ...contentContainerStyle,
-                        top: contentTopAttribute,
-                    }}
-                    key={`tabContent-${idx}`}
-                    className={`${styles.tabContent} ${
-                        activeTab === idx ? styles.show : styles.hide
-                    }`}
-                >
-                    {content}
-                </div>
-            ))}
+            <div
+                style={{
+                    ...contentContainerStyle,
+                    top: contentTopAttribute,
+                }}
+                className={styles.tabContent}
+            >
+                {content}
+            </div>
         </div>
     );
 }
