+++ src/data/constants/constants.ts
@@ -4,7 +4,6 @@ export const LOCAL_STORAGE_TOKEN_KEY = "tokens";
 //Routes
 export const GUEST_HOME_ROUTE = "/guest/";
 export const HOME_ROUTE = "/";
-
 export const ADMIN_ROUTE = "/admin/";
 export const LOGIN_REGISTER_ROUTE = "/login/";
 export const EDIT_QUESTION_ROUTE = "/questions/edit/";
