+++ src/locale/en.json
@@ -115,6 +115,7 @@
         "delete": "Delete",
         "edit": "Edit",
         "couldNotFindQuestion": "Oops! We could not find this question",
-        "editQuestion": "Edit your question"
+        "editQuestion": "Edit your question",
+        "saveChanges": "Save changes"
     }
 }
