+++ src/pages/edit/EditQuestion.tsx
@@ -1,8 +1,7 @@
 import { useParams } from "react-router";
-import { EditAQuestion, RouteButton } from "../../components";
+import { EditAQuestion, GoBackButton } from "../../components";
 import { useTranslation } from "react-i18next";
 import styles from "./EditQuestion.module.css";
-import { HOME_ROUTE } from "../../data";
 
 export function EditQuestion() {
     const { questionId } = useParams();
@@ -15,10 +14,7 @@ export function EditQuestion() {
     return (
         <section className={styles.container}>
             <div className={styles.alignleft}>
-                <RouteButton
-                    routeTo={HOME_ROUTE}
-                    text={t("backToQA")}
-                />
+                <GoBackButton text={t("backToQA")} />
             </div>
 
             <h2>{t("editQuestion")}</h2>
